split plan node representation into a crate
