+++ datafusion-optd-cli/Cargo.toml
@@ -57,7 +57,7 @@ tokio = { version = "1.24", features = [
     "parking_lot",
 ] }
 url = "2.2"
-datafusion-optd-bridge = { path = "../datafusion-optd-bridge" }
+optd-datafusion-bridge = { path = "../optd-datafusion-bridge" }
 
 [dev-dependencies]
 assert_cmd = "2.0"
