+++ optd-core/src/lib.rs
@@ -1,5 +1,4 @@
 pub mod cascades;
 pub mod cost;
-pub mod plan_nodes;
 pub mod rel_node;
 pub mod rules;
