+++ optd-datafusion-repr/Cargo.toml
@@ -1,0 +1,17 @@
+[package]
+name = "optd-datafusion-repr"
+version = "0.1.0"
+edition = "2021"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+anyhow = "1"
+num-traits = "0.2"
+num-derive = "0.2"
+tracing = "0.1"
+ordered-float = "4"
+tracing-subscriber = "0.3"
+pretty-xmlish = "0.1"
+itertools = "0.11"
+optd-core = { path = "../optd-core" }
