+++ optd-datafusion-repr/src/plan_nodes/join.rs
@@ -3,7 +3,7 @@ use std::fmt::Display;
 
 use pretty_xmlish::Pretty;
 
-use crate::rel_node::RelNode;
+use optd_core::rel_node::RelNode;
 
 use super::{replace_typ, Expr, OptRelNode, OptRelNodeRef, OptRelNodeTyp, PlanNode};
