+++ optd-datafusion-repr/src/rules.rs
@@ -1,0 +1,9 @@
+mod filter_join;
+mod join_assoc;
+mod join_commute;
+mod physical;
+
+pub use filter_join::FilterJoinPullUpRule;
+pub use join_assoc::{JoinAssocLeftRule, JoinAssocRightRule};
+pub use join_commute::JoinCommuteRule;
+pub use physical::PhysicalConversionRule;
