+++ optd-datafusion-repr/src/rules/join_assoc.rs
@@ -1,14 +1,9 @@
 use std::collections::HashMap;
 
-use crate::{
-    plan_nodes::{JoinType, OptRelNodeTyp},
-    rel_node::RelNode,
-};
+use optd_core::rel_node::RelNode;
+use optd_core::rules::{OneOrMany, Rule, RuleMatcher};
 
-use super::{
-    ir::{OneOrMany, RuleMatcher},
-    Rule,
-};
+use crate::plan_nodes::{JoinType, OptRelNodeTyp};
 
 /// Implements A join (B join C) = (A join B) join C
 pub struct JoinAssocLeftRule {
