+++ optd-datafusion-repr/src/rules/join_commute.rs
@@ -1,14 +1,9 @@
 use std::collections::HashMap;
 
-use crate::{
-    plan_nodes::{JoinType, OptRelNodeTyp},
-    rel_node::RelNode,
-};
+use optd_core::rel_node::RelNode;
+use optd_core::rules::{OneOrMany, Rule, RuleMatcher};
 
-use super::{
-    ir::{OneOrMany, RuleMatcher},
-    Rule,
-};
+use crate::plan_nodes::{JoinType, OptRelNodeTyp};
 
 /// Implements A join B = B join A
 /// TODO: should insert a projection to reorder the columns
